refactor(core): separated types into their own modules
