+++ harper-core/src/linting/mod.rs
@@ -15,6 +15,7 @@ mod lets_confusion;
 mod linking_verbs;
 mod lint;
 mod lint_group;
+mod lint_kind;
 mod long_sentences;
 mod matcher;
 mod merge_linters;
@@ -32,6 +33,7 @@ mod somewhat_something;
 mod spaces;
 mod spell_check;
 mod spelled_numbers;
+mod suggestion;
 mod terminating_conjunctions;
 mod that_which;
 mod unclosed_quotes;
@@ -48,8 +50,9 @@ pub use dot_initialisms::DotInitialisms;
 pub use ellipsis_length::EllipsisLength;
 pub use lets_confusion::LetsConfusion;
 pub use linking_verbs::LinkingVerbs;
-pub use lint::{Lint, LintKind, Suggestion};
+pub use lint::Lint;
 pub use lint_group::{LintGroup, LintGroupConfig};
+pub use lint_kind::LintKind;
 pub use long_sentences::LongSentences;
 pub use matcher::Matcher;
 pub use merge_words::MergeWords;
@@ -69,6 +72,7 @@ pub use somewhat_something::SomewhatSomething;
 pub use spaces::Spaces;
 pub use spell_check::SpellCheck;
 pub use spelled_numbers::SpelledNumbers;
+pub use suggestion::Suggestion;
 pub use terminating_conjunctions::TerminatingConjunctions;
 pub use that_which::ThatWhich;
 pub use unclosed_quotes::UnclosedQuotes;
